fix(core): pack/unpack fp16 for the EdgeTAM video pipeline
The video graphs (videoEncoder/memoryAttention/maskDecoderVideo/
memoryEncoder) are exported entirely in float16, but VideoEngine and
MemoryBank uploaded/allocated every host tensor as float32, and read
model outputs back as if they were float32 — causing ORT to reject
inputs ("expected tensor(float16), actual tensor(float)") and, further
downstream, the memory-bank ring's copyRegion to hard-fail on a
src/dst dtype mismatch against memoryEncoder's fp16 outputs.

Adds packages/core/src/runtime/float16.ts (no such conversion existed
in the repo) and packs every host-generated tensor to half bits before
upload, and unpacks fp16 readbacks (mask logits, IoU/occlusion scores,
object pointer) back to real floats before any numeric postprocessing
— fixing both the ORT crash and what would otherwise be silent mask
corruption from mis-decoded raw fp16 bits.

+++ packages/core/src/worker/video/memory-bank.test.ts
@@ -142,10 +142,17 @@ function makeBank(backend: FakeBackend, overrides?: Partial<VideoManifestSection
   return new MemoryBank({ backend, video: v, strategy: strategyFor('edgetam', v), location: 'device' });
 }
 
-/** Sentinel feature/pos tensors filled with `value` so slot writes are checkable. */
+/**
+ * Sentinel feature/pos tensors filled with `value` so slot writes are
+ * checkable. Tagged `'float16'` — the ring `MemoryBank` allocates is
+ * `'float16'` (the real memoryEncoder outputs it commits are), and
+ * `copyRegion` requires src/dst dtypes to match; the fake backend's `data`
+ * stays a plain `Float32Array` regardless since this harness never encodes
+ * real half-float bits.
+ */
 function feat(backend: FakeBackend, value: number, shape: readonly number[] = [T, MEM_DIM]): DeviceTensor {
   const data = new Float32Array(elemCount(shape)).fill(value);
-  return backend.uploadTensor(data, shape, 'float32');
+  return backend.uploadTensor(data, shape, 'float16');
 }
 
 function pointerVec(value: number): Float32Array {
